@@ -19,34 +19,35 @@
     "publish": "npm run build-public && node utils/github-publish"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "autoprefixer": "^9.6.0",
-    "babel-eslint": "^10.0.2",
-    "babel-loader": "^8.0.6",
-    "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^3.0.0",
-    "eslint": "^5.16.0",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-react": "^7.11.1",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "gh-pages": "^2.0.1",
-    "html-webpack-plugin": "^3.2.0",
-    "postcss-loader": "^3.0.0",
-    "prop-types": "^15.6.2",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.16.5",
-    "webpack-cli": "^3.1.0",
-    "webpack-dashboard": "^3.0.7",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.1"
+    "@babel/cli": "^7.27.2",
+    "@babel/core": "^7.27.4",
+    "@babel/eslint-parser": "^7.27.5",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.27.4",
+    "@babel/preset-env": "^7.27.2",
+    "@babel/preset-react": "^7.27.1",
+    "autoprefixer": "^10.4.21",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^9.1.3",
+    "copy-webpack-plugin": "^12.0.0",
+    "css-loader": "^6.8.1",
+    "eslint": "^9.29.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-webpack-plugin": "^4.0.1",
+    "gh-pages": "^6.3.0",
+    "html-webpack-plugin": "^5.5.3",
+    "mini-css-extract-plugin": "^2.9.2",
+    "postcss-loader": "^7.3.3",
+    "prop-types": "^15.8.1",
+    "style-loader": "^3.3.4",
+    "webpack": "^5.99.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dashboard": "^3.3.8",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.10.0"
   },
   "dependencies": {
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
   }
 }
