@@ -1,14 +1,13 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import { Player } from '../src/'
 
 const root = document.getElementById('webvtt-player')
 
-ReactDOM.render(
+ReactDOM.createRoot(root).render(
   <Player
     audio={root.dataset.audio}
     transcript={root.dataset.transcript}
     metadata={root.dataset.metadata}
-    preload={true} />,
-  root
+    preload={true} />
 )
